+++ temba/channels/handlers.py
@@ -1153,7 +1153,7 @@ class M3TechHandler(ExternalHandler):
     handler_name = 'handlers.m3tech_handler'
 
     def get_channel_type(self):
-        return Channel.TYPE_M3TECH
+        return 'M3'
 
 
 class NexmoCallHandler(BaseChannelHandler):
@@ -1532,7 +1532,7 @@ class KannelHandler(BaseChannelHandler):
         request_uuid = kwargs['uuid']
 
         # look up the channel
-        channel = Channel.objects.filter(uuid=request_uuid, is_active=True, channel_type=Channel.TYPE_KANNEL).exclude(org=None).first()
+        channel = Channel.objects.filter(uuid=request_uuid, is_active=True, channel_type='KN').exclude(org=None).first()
         if not channel:
             return HttpResponse("Channel not found for id: %s" % request_uuid, status=400)
 
@@ -2116,7 +2116,7 @@ class JunebugHandler(BaseChannelHandler):
         data = json.load(request)
         is_ussd = self.is_ussd_message(data)
         channel_data = data.get('channel_data', {})
-        channel_types = (Channel.TYPE_JUNEBUG_USSD, Channel.TYPE_JUNEBUG)
+        channel_types = ('JNU', 'JN')
 
         # look up the channel
         channel = Channel.objects.filter(