+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+v3.0.209
+----------
+ * Add a way to specify the prefixes short codes should be matching
+ * Include both high_priority and priority in courier JSON
+ * Fix TwiML migration
+ * Fix JSON response when searching Plivo numbers
+
 v3.0.208
 ----------
  * Msg.bulk_priority -> Msg.high_priority